requirements.txt
open3d==0.19.0
matplotlib
opencv-python
PySide6
PyOpenGL==3.1.6
PyGLM>=2.5.0